Добавлен Rfc2898Encryptor.
