+++ MailSender.lib/Interface/IEncryptorService.cs
@@ -1,0 +1,11 @@
+﻿namespace MailSender.lib.Interface
+{
+    public interface IEncryptorService
+    {
+        string Encrypt(string str, string password);
+        string Decrypt(string str, string password);
+
+        byte[] Encrypt(byte[] data, string password);
+        byte[] Decrypt(byte[] data, string password);
+    }
+}
