+++ MailSender.lib/Services/TextEncoder.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using MailSender.lib.Interface;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
